Refactor: 프로필 리팩토링 중. 폴더구조 변경함
