--- src/components/OwnerProject/Card/Card.constants.ts
@@ -1,1 +1,0 @@
-export const CARD_MODAL = "CARD_MODAL";
