--- src/components/ProjectCreate/DatePicker/DatePicker.constants.ts
@@ -1,1 +1,0 @@
-export const DATE_PICKER_MODAL = "DATE_PICKER_MODAL";
