--- src/components/ProjectCreate/DatePicker/DatePicker.types.ts
@@ -1,3 +1,0 @@
-export interface DatePickerProps {
-  onChangeDate?: (date: { startDate: Date | null; endDate: Date | null }) => void;
-}
