--- src/components/ProjectCreate/DatePicker/index.ts
@@ -1,2 +1,0 @@
-export { DatePicker } from "./DatePicker";
-export { DATE_PICKER_MODAL } from "./DatePicker.constants";
