--- src/components/ProjectCreate/index.ts
@@ -1,1 +1,0 @@
-export { DatePicker } from "./DatePicker";
