+++ src/pages/profile/component/SectionTab/SectionTab.styles.ts
