+++ src/pages/profile/component/SectionTab/SectionTab.types.ts
