+++ src/pages/profile/index.page.tsx
@@ -4,9 +4,9 @@ import { serverSideTranslations } from "next-i18next/serverSideTranslations";
 import Link from "next/link";
 import { useRouter } from "next/router";
 import { IconButton } from "~/components/Commons";
-import { Member, Owner } from "~/components/Profile";
 import { Navbar } from "~/components/Shared";
 import { routes } from "~/constants/routes";
+import { Member, Owner } from "~/pages/profile/component";
 import { profileKeys, selectProfile, useSelectProfileQuery } from "~/states/server/profile";
 import { LayoutContent, LayoutHeaderWithNav } from "~/styles/mixins";
 import { requireAuthentication } from "~/utils";
